+++ src/utils/MainApi.js
@@ -1,8 +1,6 @@
 const BASE_URL = "http://localhost:3001";
 // const BASE_URL = "https://api.movies-favorite.nomoredomains.work";
 
-const BASE_URL_MOVIES_API = "https://api.nomoreparties.co";
-
 class MainApi {
   constructor(url) {
     this._url = url;
@@ -56,19 +54,7 @@ class MainApi {
       method: "POST",
       headers: this._headers,
       credentials: this._credentials,
-      body: JSON.stringify({
-        country: data.country,
-        director: data.director,
-        duration: data.duration,
-        year: data.year,
-        description: data.description,
-        image: BASE_URL_MOVIES_API + data.image.url,
-        trailerLink: data.trailerLink,
-        nameRU: data.nameRU,
-        nameEN: data.nameEN,
-        thumbnail: BASE_URL_MOVIES_API + data.image.formats.thumbnail.url,
-        movieId: data.id,
-      }),
+      body: JSON.stringify(data),
     }).then(this._checkResult);
   }
